+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
+- Return a pipeline summary model with more information from the Pipeline ExecutePipelineAsync call
 - Fixed race condition with trx file parsing before the file was saved for DotNet Test